% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmconvert.R
\name{osmt_convert}
\alias{osmt_convert}
\title{Convert file format of osm files}
\usage{
osmt_convert(file, format_out = "o5m", drop_author = TRUE)
}
\description{
@param file path to file
 @param format format to convert to
}
