% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmfilter.R
\name{osmt_filter}
\alias{osmt_filter}
\title{Filter osm by tags}
\usage{
osmt_filter(file, path_out, keep = NULL, drop = NULL,
  keep_tags = NULL, drop_tags = NULL)
}
\description{
@param file path to file
 @param path_out path to file to save result
 @param keep object to keep arguments
 @param drop object to drop arguments
 @param keep_tags tags to keep arguments
 @param drop_tags togs to drop arguments
}
