% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofabrik.R
\name{osmt_geofabrik_dl}
\alias{osmt_geofabrik_dl}
\title{Download OSM from Geofabrik}
\usage{

  osmt_geofabrik_dl(region = "europe/great-britain/england/west-yorkshire",
  destfile = "osm.pbf")
}
\arguments{
\item{region}{geogabrik path to file without start and end e.g. "europe/great-britain/england/west-yorkshire"}

\item{destfile}{file path to save file to}
}
\description{
Download OSM from Geofabrik
}
