% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmconvert.R
\name{osmt_convert}
\alias{osmt_convert}
\title{Convert file format of osm files}
\usage{
osmt_convert(file, format_out = "pbf", drop_author = TRUE, overwrite = FALSE)
}
\arguments{
\item{file}{path to file}

\item{format_out}{format to convert to}


\item{drop_author}{logical, reduce file size bu removing author info}

\item{overwrite}{logical, should existing file be overwritten, default = FALSE}

\description{
Convert file format of osm files
}
