% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmfilter.R
\name{osmt_filter}
\alias{osmt_filter}
\title{Filter osm by tags}
\usage{
osmt_filter(file, path_out, keep = NULL, drop = NULL,
  keep_tags = NULL, drop_tags = NULL)
}
\arguments{
\item{file}{path to file}

\item{path_out}{path to file to save result}

\item{keep}{object to keep arguments}

\item{drop}{object to drop arguments}

\item{keep_tags}{tags to keep arguments}

\item{drop_tags}{togs to drop arguments}
}
\description{
Filter osm by tags
}
