% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmfilter.R
\name{osmt_filter}
\alias{osmt_filter}
\title{Filter osm by tags}
\usage{
osmt_filter(
  file,
  path_out,
  keep = NULL,
  drop = NULL,
  modify = NULL,
  keep_tags = NULL,
  drop_tags = NULL,
  modify_tags = NULL,
  bbox = NULL
)
}
\arguments{
\item{file}{path to file}

\item{path_out}{path to file to save result}

\item{keep}{character, object to keep arguments}

\item{drop}{character, object to drop arguments}

\item{modify}{character, object tags paris to modify}

\item{keep_tags}{character, tags to keep arguments}

\item{drop_tags}{character, tags to drop arguments}

\item{modify_tags}{character, tags to modify}

\item{bbox}{sf bbox, from st_bbox(), or a 2x2 matrix lng/lat coordinates for bottom left and top right corners}
}
\value{
FUnction returns NULL but a new file is saved to disk
}
\description{
https://wiki.openstreetmap.org/wiki/Osmfilter
}
\details{
A wrapper around osmfilter, supports Windows, Mac, Linux versions
}
